+++ filename_unix/src/filename_unix_stubs.c
@@ -28,7 +28,7 @@ CAMLprim value core_unix_realpath(value v_path) {
 }
 #else
 CAMLprim value core_unix_realpath(value v_path) {
-  char *path = String_val(v_path);
+  const char *path = String_val(v_path);
   /* [realpath] is inherently broken without GNU-extension, and this
      seems like a reasonable thing to do if we do not build against
      GLIBC. */