+++ linux_ext/src/linux_ext_intf.ml
@@ -349,7 +349,7 @@ module type S = sig
 
     (*_ See the Jane Street Style Guide for an explanation of [Private] submodules:
 
-      https://opensource.janestreet.com/standards/#private-submodules *)
+        https://opensource.janestreet.com/standards/#private-submodules *)
     module Private : sig
       val unsafe_timerfd_settime
         :  File_descr.t